requirements.txt
aplpy==2.1.0
asciitree==0.3.3
asgiref==3.8.1
astropy==6.1.2
astropy-iers-data==0.2024.8.12.0.32.58
astropy_healpix==1.0.3
astroquery==0.4.7
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
dask==2024.8.0
Django==5.1
fasteners==0.19
fonttools==4.53.1
fsspec==2024.6.1
greenlet==3.0.3
html5lib==1.1
idna==3.7
imageio==2.34.2
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.2
jeepney==0.8.0
jplephem==2.22
keyring==25.3.0
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.43.0
locket==1.0.0
matplotlib==3.9.2
miepython==2.5.4
more-itertools==10.4.0
networkx==3.3
numba==0.60.0
numcodecs==0.13.0
numpy==2.0.1
packaging==24.1
pandas==2.2.2
partd==1.4.2
pillow==10.4.0
PyAVM==0.9.6
pycparser==2.22
pyerfa==2.0.1.4
pyparsing==3.1.2
pyregion==2.2.0
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2024.1
pyvo==1.5.2
PyYAML==6.0.2
query==0.1.4
reproject==0.14.0
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.0
SecretStorage==3.3.3
sgp4==2.23
shapely==2.0.5
six==1.16.0
skyfield==1.49
soupsieve==2.6
SQLAlchemy==2.0.32
sqlparse==0.5.1
tifffile==2024.8.10
toolz==0.12.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
utils==1.0.2
webencodings==0.5.1
zarr==2.18.2
